fix layout padding
