+++ src/app/(sidebar)/account/create/page.tsx
@@ -1,14 +1,26 @@
 "use client";
 import { useRouter } from "next/navigation";
 import { Card, Text, Button } from "@stellar/design-system";
+import { Keypair } from "stellar-sdk";
 
 import { GenerateKeypair } from "@/components/GenerateKeypair";
 import { Routes } from "@/constants/routes";
+import { useStore } from "@/store/useStore";
 
 import "../styles.scss";
 
 export default function CreateAccount() {
+  const { account } = useStore();
   const router = useRouter();
+
+  const generateKeypair = () => {
+    let keypair = Keypair.random();
+
+    account.updateKeypair({
+      publicKey: keypair.publicKey(),
+      secretKey: keypair.secret(),
+    });
+  };
 
   return (
     <div className="Account">
@@ -26,7 +38,9 @@ export default function CreateAccount() {
             </Text>
           </div>
           <div className="Account__CTA">
-            <GenerateKeypair />
+            <Button size="md" variant="tertiary" onClick={generateKeypair}>
+              Generate keypair
+            </Button>
 
             <Button
               size="md"
@@ -36,6 +50,8 @@ export default function CreateAccount() {
               Fund account with Friendbot
             </Button>
           </div>
+
+          <GenerateKeypair />
         </div>
       </Card>
     </div>
